STUDY_NEW/requirements.txt
flask==2.0.1
flask-session==0.4.0
gunicorn==20.1.0
bcrypt==4.0.1
psycopg2-binary==2.9.6